improve windowborder element
