+++ src/quick/framelesshelperquick_global.cpp
@@ -30,8 +30,8 @@
 #  include "quickchromepalette.h"
 #  include "quickstandardsystembutton_p.h"
 #  include "quickstandardtitlebar_p.h"
+#  include "framelessquickwindow.h"
 #  include "framelessquickwindow_p.h"
-#  include "framelessquickwindow_p_p.h"
 #  include "quickmicamaterial.h"
 #  include "quickmicamaterial_p.h"
 #  include "quickimageitem.h"
